UC302 service creation
Made a dynamic service for updating a meal:

First it says the allowedfields that can be updated.

Then for each allowed field it push the field with the data. When you want to update the allergenes, then it will set the allergenes field as a comma-separated string.

Then it activates the query and tada, results